Extract GTModelMaterial test method to class
