+++ src/test/java/org/opengis/cite/cdb10/cdbStructure/VerifyGTModelStructureTests.java
@@ -960,103 +960,4 @@ public class VerifyGTModelStructureTests extends StructureTestFixture<GTModelStr
 		// execute
 		this.testSuite.verifyModelTextureFile();
 	}
-	
-	
-	/*
-	 * Verify Model Material Files
-	 */
-	@Test(expected = AssertionError.class)
-	public void verifyModelMaterialFile_invalid() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"asdf")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test(expected = AssertionError.class)
-	public void verifyModelMaterialFile_invalidDataset() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"D211_S001_T001_L10_AC_1.tif")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test(expected = AssertionError.class)
-	public void verifyModelMaterialFile_invalidExtension() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"D504_S001_T001_L10_AC_1.txt")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test(expected = AssertionError.class)
-	public void verifyModelMaterialFile_badCS1Length() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"D504_S1_T001_L10_AC_1.tif")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test(expected = AssertionError.class)
-	public void verifyModelMaterialFile_badCS1Type() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"D504_SABC_T001_L10_AC_1.tif")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test(expected = AssertionError.class)
-	public void verifyModelMaterialFile_badCS2Length() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"D504_S001_T1_L10_AC_1.tif")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test(expected = AssertionError.class)
-	public void verifyModelMaterialFile_badCS2Type() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"D504_S001_TABC_L10_AC_1.tif")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test
-	public void verifyModelMaterialFile_Good() throws IOException {
-		// setup
-		Files.createDirectories(this.cdb_root.resolve(Paths.get("GTModel", "504_GTModelMaterial",
-				"A", "B", "AB_1",
-				"D504_S001_T001_L10_AB_1.tif")));
-
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
-
-	@Test
-	public void verifyModelMaterialFile_Skip() throws IOException {
-		// setup
-		// execute
-		this.testSuite.verifyModelMaterialFile();
-	}
 }
